cilium: improve help dump again
This reverts commit 06389872ceb5 ("Improve 'cilium-agent --help'") which
overall makes the --help output less helpful than before. There are various
issues, e.g. flags are forgotten to be added to sections given this needs
to be explicitly specified and then we end up with a random chunk of help
flags at the end of the --help dump.

The sectioning is not overly useful and various flags end up in multiple
sections or in less fitting sections and overall this leaves (imho) more
confusion to a user than what it is supposed to resolve.

Given the complex setup possibilities in Cilium best way might be from
getting started guide or cmdref (which I rather tend to end up looking than
current --help), and an alphabetical sorting of options is (imho) still
better and easier than to search multiple sections of flags, hence make
output similar to cmdref. Overall as also discussed with others this would
need an improved approach than 06389872ceb5 once we revisit this again.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>